MODI: DetailViewController - 게시글 상세 화면 수정 #11
