+++ DongJibSa/DongJibSa/Controller/Home/DetailViewController+TableViewController.swift
@@ -8,34 +8,62 @@
 import UIKit
 
 extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 4
+    }
+    
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return table
+        if section == 0 {
+            return 1
+        } else if section == 2{
+            return table
+        } else if section == 3 {
+            return 0
+        } else {
+            return 3
+        }
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        if tableView == ingredientsTableView {
-            let cell = tableView.dequeueReusableCell(withIdentifier: IngredientsInfoCell.cellID, for: indexPath) as! IngredientsInfoCell
+        switch indexPath.section {
+        case 0:
+            let cell = tableView.dequeueReusableCell(withIdentifier: RecipeInfoCell.cellId, for: indexPath) as! RecipeInfoCell
+            cell.selectionStyle = .none
             return cell
-        } else {
+        case 1:
+            let cell = tableView.dequeueReusableCell(withIdentifier: IngredientsInfoCell.cellId, for: indexPath) as! IngredientsInfoCell
+            cell.selectionStyle = .none
+            return cell
+        case 2:
             let cell = tableView.dequeueReusableCell(withIdentifier: CommentCell.cellId, for: indexPath) as! CommentCell
+            cell.selectionStyle = .none
+            return cell
+        default:
+            let cell = tableView.dequeueReusableCell(withIdentifier: EmptyCell.cellId, for: indexPath) as! EmptyCell
+            cell.selectionStyle = .none
             return cell
         }
     }
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        if tableView == ingredientsTableView {
+        if indexPath.section == 0 {
+          return 250
+        } else if indexPath.section == 1 {
             return 44
-        } else {
+        } else if indexPath.section == 2 {
             return 100
+        } else {
+            return 0
         }
     }
     
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        if tableView == ingredientsTableView {
+        if section == 1 {
             return 44
+        } else if section == 3 {
+            return 50
         } else {
-            return 30
+            return 65
         }
     }
     
@@ -43,8 +71,54 @@ extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
         let headerView = UIView()
         headerView.backgroundColor = .white
         
-        if tableView == ingredientsTableView {
+        if section == 0 {
+            let profileImageView: UIImageView = {
+                let imageView = UIImageView()
+                imageView.backgroundColor = .primaryColor
+                imageView.layer.cornerRadius = 45 / 2
+                return imageView
+            }()
             
+            headerView.addSubview(profileImageView)
+            profileImageView.snp.makeConstraints { make in
+                make.centerY.equalToSuperview()
+                make.left.equalToSuperview().inset(16)
+                make.width.height.equalTo(45)
+            }
+            
+            let userNameLabel: UILabel = {
+                let label = UILabel()
+                label.text = "집밥이지"
+                label.font = .boldSystemFont(ofSize: 16)
+                return label
+            }()
+            
+            let descriptionLabel: UILabel = {
+                let label = UILabel()
+                label.text = "신대방동"
+                label.font = .systemFont(ofSize: 14)
+                return label
+            }()
+            
+            let profileStackView: UIStackView = {
+                let stackView = UIStackView(arrangedSubviews: [userNameLabel, descriptionLabel])
+                stackView.axis = .vertical
+                stackView.distribution = .fillEqually
+                stackView.alignment = .fill
+                stackView.spacing = 3
+                return stackView
+            }()
+            
+            headerView.addSubview(profileStackView)
+            profileStackView.snp.makeConstraints { make in
+                make.centerY.equalTo(profileImageView.snp.centerY)
+                make.left.equalTo(profileImageView.snp.right).offset(10)
+                make.right.equalToSuperview().inset(16)
+            }
+            
+            return headerView
+            
+        } else if section == 1 {
             let titleLabel = UILabel()
             titleLabel.text = "재료명"
             titleLabel.font = .systemFont(ofSize: 14)
@@ -78,7 +152,8 @@ extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
             
             headerView.addSubview(stackView)
             stackView.snp.makeConstraints { make in
-                make.edges.equalToSuperview()
+                make.center.equalToSuperview()
+                make.left.right.equalToSuperview().inset(16)
             }
             
             let seperateView = UIView()
@@ -88,18 +163,116 @@ extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
             seperateView.snp.makeConstraints { make in
                 make.bottom.equalToSuperview()
                 make.height.equalTo(1)
-                make.left.right.equalToSuperview()
+                make.left.right.equalToSuperview().inset(16)
             }
             
             return headerView
-        } else {
-//            let registeredImageView: UIImageView = {
-//                let imageView = UIImageView()
-//                imageView.image = UIImage(systemName: "circle.fill")
-//                imageView.tintColor = .primaryColor
-//                imageView.contentMode = .scaleAspectFit
-//                return imageView
-//            }()
+            
+        } else if section == 2 {
+            let likeImageView: UIImageView = {
+                let imageView = UIImageView()
+                imageView.image = UIImage(systemName: "heart.fill")
+                imageView.tintColor = .gray
+                return imageView
+            }()
+            
+            let likeLabel: UILabel = {
+                let label = UILabel()
+                label.text = "7"
+                label.textColor = .gray
+                label.font = .systemFont(ofSize: 12)
+                return label
+            }()
+            
+            let likeStackView: UIStackView = {
+                let stackView = UIStackView(arrangedSubviews: [likeImageView, likeLabel])
+                stackView.axis = .horizontal
+                stackView.distribution = .fill
+                stackView.alignment = .fill
+                stackView.spacing = 8
+                return stackView
+            }()
+            
+            let talkImageView: UIImageView = {
+                let imageView = UIImageView()
+                imageView.image = UIImage(systemName: "message")
+                imageView.tintColor = .gray
+                return imageView
+            }()
+            
+            let talkLabel: UILabel = {
+                let label = UILabel()
+                label.text = "3"
+                label.textColor = .gray
+                label.font = .systemFont(ofSize: 12)
+                return label
+            }()
+            
+            let talkStackView: UIStackView = {
+                let stackView = UIStackView(arrangedSubviews: [talkImageView, talkLabel])
+                stackView.axis = .horizontal
+                stackView.distribution = .fill
+                stackView.alignment = .fill
+                stackView.spacing = 8
+                return stackView
+            }()
+            
+            let viewImageView: UIImageView = {
+                let imageView = UIImageView()
+                imageView.image = UIImage(systemName: "eye")
+                imageView.tintColor = .gray
+                return imageView
+            }()
+            
+            let viewLabel: UILabel = {
+                let label = UILabel()
+                label.text = "24"
+                label.textColor = .gray
+                label.font = .systemFont(ofSize: 12)
+                return label
+            }()
+            
+            let viewStackView: UIStackView = {
+                let stackView = UIStackView(arrangedSubviews: [viewImageView, viewLabel])
+                stackView.axis = .horizontal
+                stackView.distribution = .fill
+                stackView.alignment = .fill
+                stackView.spacing = 8
+                return stackView
+            }()
+            
+            let emptyView: UIView = {
+                let view = UIView()
+                return view
+            }()
+            
+            let listStackView: UIStackView = {
+                let stackView = UIStackView(arrangedSubviews: [likeStackView, talkStackView, emptyView, viewStackView])
+                stackView.axis = .horizontal
+                stackView.distribution = .fill
+                stackView.alignment = .fill
+                stackView.spacing = 20
+                return stackView
+            }()
+            
+            headerView.addSubview(listStackView)
+            listStackView.snp.makeConstraints { make in
+                make.top.equalToSuperview().offset(4)
+                make.left.right.equalToSuperview().inset(16)
+            }
+            
+            let seperateView: UIView = {
+                let view = UIView()
+                view.backgroundColor = .systemGray3
+                return view
+            }()
+            
+            headerView.addSubview(seperateView)
+            seperateView.snp.makeConstraints { make in
+                make.top.equalTo(listStackView.snp.bottom).offset(5)
+                make.left.right.equalToSuperview().inset(16)
+                make.height.equalTo(1)
+            }
             
             let registeredButton: UIButton = {
                 let button = UIButton()
@@ -109,19 +282,6 @@ extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
                 return button
             }()
             
-//            let registeredStackView = UIStackView(arrangedSubviews: [registeredImageView, registeredButton])
-//            registeredStackView.axis = .horizontal
-//            registeredStackView.distribution = .fill
-//            registeredStackView.alignment = .center
-            
-//            let recentImageView: UIImageView = {
-//                let imageView = UIImageView()
-//                imageView.image = UIImage(systemName: "circle.fill")
-//                imageView.tintColor = .primaryColor
-//                imageView.contentMode = .scaleAspectFit
-//                return imageView
-//            }()
-            
             let recentButton: UIButton = {
                 let button = UIButton()
                 button.setTitle("● 최신순", for: .normal)
@@ -130,14 +290,9 @@ extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
                 return button
             }()
             
-//            let recentStackView = UIStackView(arrangedSubviews: [recentImageView, recentButton])
-//            recentStackView.axis = .horizontal
-//            recentStackView.distribution = .fill
-//            recentStackView.alignment = .center
+            let emptyView2 = UIView()
             
-            let emptyView = UIView()
-            
-            let stackView = UIStackView(arrangedSubviews: [registeredButton, recentButton, emptyView])
+            let stackView = UIStackView(arrangedSubviews: [registeredButton, recentButton, emptyView2])
             stackView.axis = .horizontal
             stackView.distribution = .fill
             stackView.alignment = .center
@@ -145,9 +300,55 @@ extension DetailViewController: UITableViewDelegate, UITableViewDataSource {
             
             headerView.addSubview(stackView)
             stackView.snp.makeConstraints { make in
-                make.edges.equalToSuperview()
+                make.top.equalTo(seperateView.snp.bottom).offset(5)
+                make.left.right.equalToSuperview().inset(16)
             }
             
+            return headerView
+            
+        } else {
+            let photoButton: UIButton = {
+                let button = UIButton()
+                button.setImage(UIImage(systemName: "photo"), for: .normal)
+                button.tintColor = .gray
+                return button
+            }()
+            
+            self.commentTextField.snp.makeConstraints { make in
+                make.height.equalTo(38)
+            }
+            
+            let sendButton: UIButton = {
+                let button = UIButton()
+                button.setImage(UIImage(systemName: "paperplane"), for: .normal)
+                button.tintColor = .gray
+                return button
+            }()
+            
+            sendButton.addTarget(self, action: #selector(addButtonTapped), for: .touchUpInside)
+            
+            let commentStackView: UIStackView = {
+                let stackView = UIStackView(arrangedSubviews: [photoButton, commentTextField, sendButton])
+                stackView.axis = .horizontal
+                stackView.distribution = .fill
+                stackView.alignment = .fill
+                stackView.spacing = 10
+                stackView.backgroundColor = .white
+                return stackView
+            }()
+
+            headerView.addSubview(commentStackView)
+            commentStackView.snp.makeConstraints { make in
+                make.center.equalToSuperview()
+                make.bottom.equalTo(headerView.keyboardLayoutGuide.snp.top)
+                make.left.right.equalToSuperview().inset(16)
+            }
+            photoButton.setContentHuggingPriority(.init(251), for: .horizontal)
+            photoButton.setContentCompressionResistancePriority(.init(751), for: .horizontal)
+            commentTextField.setContentHuggingPriority(.init(250), for: .horizontal)
+            commentTextField.setContentCompressionResistancePriority(.init(750), for: .horizontal)
+            sendButton.setContentHuggingPriority(.init(251), for: .horizontal)
+            sendButton.setContentCompressionResistancePriority(.init(751), for: .horizontal)
             return headerView
         }
