+++ DongJibSa/DongJibSa/View/DongJibSa/View/Base.lproj/Detail.storyboard
@@ -24,18 +24,27 @@
                                     <constraint firstAttribute="height" constant="375" id="mxD-lZ-dmV"/>
                                 </constraints>
                             </view>
+                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="grouped" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="18" estimatedSectionHeaderHeight="-1" sectionFooterHeight="18" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="5Y2-Be-ThX">
+                                <rect key="frame" x="0.0" y="375" width="393" height="443"/>
+                                <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                            </tableView>
                         </subviews>
                         <viewLayoutGuide key="safeArea" id="JN5-vc-fmF"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
                             <constraint firstItem="4cU-b8-Plc" firstAttribute="top" secondItem="aba-j2-DQf" secondAttribute="top" id="BJC-UR-2Zj"/>
+                            <constraint firstItem="JN5-vc-fmF" firstAttribute="bottom" secondItem="5Y2-Be-ThX" secondAttribute="bottom" id="BzJ-O6-T3c"/>
+                            <constraint firstItem="5Y2-Be-ThX" firstAttribute="top" secondItem="4cU-b8-Plc" secondAttribute="bottom" id="ZjY-mT-z5M"/>
                             <constraint firstItem="4cU-b8-Plc" firstAttribute="leading" secondItem="JN5-vc-fmF" secondAttribute="leading" id="d5Y-sc-VkR"/>
+                            <constraint firstItem="5Y2-Be-ThX" firstAttribute="leading" secondItem="JN5-vc-fmF" secondAttribute="leading" id="eOR-8W-jGV"/>
                             <constraint firstItem="JN5-vc-fmF" firstAttribute="trailing" secondItem="4cU-b8-Plc" secondAttribute="trailing" id="tLb-6z-6MR"/>
+                            <constraint firstItem="JN5-vc-fmF" firstAttribute="trailing" secondItem="5Y2-Be-ThX" secondAttribute="trailing" id="xOv-kE-epA"/>
                         </constraints>
                     </view>
                     <connections>
                         <outlet property="recipeView" destination="4cU-b8-Plc" id="rfz-xS-6oF"/>
                         <outlet property="recipeViewHeight" destination="mxD-lZ-dmV" id="jbe-eX-FpU"/>
+                        <outlet property="tableView" destination="5Y2-Be-ThX" id="xQk-DK-2A0"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="0MK-Qo-lUi" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
